editor/libeditor/tests/test_nsITableEditor_deleteTableCell.html

<!DOCTYPE>
<html>
<head>
  <title>Test for nsITableEditor.deleteTableCell()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<div id="content" contenteditable>out of table<table><tr><td>default content</td></tr></table></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  let editor = document.getElementById("content");
  let selection = document.getSelection();

  selection.collapse(editor.firstChild, 0);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "out of table<table><tbody><tr><td>default content</td></tr></tbody></table>",
     "nsITableEditor.deleteTableCell(1) should do nothing if selection is not in <table>");

  selection.removeAllRanges();
  try {
    getTableEditor().deleteTableCell(1);
    ok(false, "getTableEditor().deleteTableCell(1) without selection ranges should throw exception");
  } catch (e) {
    ok(true, "getTableEditor().deleteTableCell(1) without selection ranges should throw exception");
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  let range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-2</td><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove only selected cell when only one cell is selected #1-1");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select">cell1-2</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove only selected cell when only one cell is selected #1-2");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td>cell1-2</td><td id="select">cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td><td>cell1-2</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove only selected cell when only one cell is selected #1-3");

  // When only one cell element is selected, the argument should be used.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(2);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(2) should remove selected cell element and next cell element in same row");

  // When the argument is larger than remaining cell elements from selected
  // cell element, the behavior is really buggy.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td>cell1-2</td><td id="select">cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(2);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-1</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(2) should remove selected cell element and its previous cell element when it reaches the last cell element in the row");

  // XXX If the former case is expected, first row should be removed in this
  //     case, but it removes only selected cell and its previous cell.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td>cell1-2</td><td id="select">cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(4);
  todo_is(editor.innerHTML, "<table><tbody>" +
                               "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(4) should remove the first row when a cell in it is selected and the argument is larger than number of cells in the row");

  // If 2 or more cells are selected, the argument should be ignored.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1">cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>' +
                       '<tr><td>cell2-1</td><td id="select2">cell2-2</td><td>cell2-3</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-2</td><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove selected cell elements even if the argument is smaller than number of selected cells");

  // If all cells in a row are selected, the <tr> element should also be removed.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1">cell1-1</td><td id="select2">cell1-2</td><td id="select3">cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select3"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove also <tr> element when all cell elements in a row is selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td></tr>' +
                       "<tr><td>cell2-1</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell2-1</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove also <tr> element when a cell element which is only child of a row is selected");

  // If all cells are removed, the <table> element should be removed.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1">cell1-1</td><td id="select2">cell1-2</td></tr>' +
                       '<tr><td id="select3">cell2-1</td><td id="select4">cell2-2</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select3"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select4"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "",
          "nsITableEditor.deleteTableCellContents(1) should remove also <table> element when all cell elements are selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableCellContents(1) should remove also <table> element when a cell element which is only child of <table> is selected");

  // rowspan
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select" rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-2</td></tr>" +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-2</td></tr>" +
                         "<tr><td>cell2-2</td></tr>" +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should just remove the cell spanning rows");

  // XXX cell3-1 is also removed even though it's not selected.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1" rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-2</td></tr>" +
                       '<tr><td>cell3-1</td><td id="select2">cell3-2</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              "<tr><td>cell1-2</td></tr>" +
                              "<tr><td>cell2-2</td></tr>" +
                              "<tr><td>cell3-1</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should just remove the cell spanning rows (when 2 cell elements are selected)");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                       '<tr><td id="select">cell2-2</td></tr>' +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td rowspan="1">cell1-1</td><td>cell1-2</td></tr>' +
                         "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should adjust rowspan when spanned <tr>'s last child cell element is removed");

  // XXX broken case, the second row isn't removed.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                       '<tr><td id="select1">cell2-2</td></tr>' +
                       '<tr><td>cell3-1</td><td id="select2">cell3-2</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              '<tr><td rowspan="1">cell1-1</td><td>cell1-2</td></tr>' +
                              "<tr><td>cell3-1</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust rowspan when spanned <tr>'s last child cell element is removed (when 2 cell elements are selected)");

  // XXX broken case, neither the selected cell nor the second row is removed.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td rowspan="3">cell1-1</td><td>cell1-2</td></tr>' +
                       '<tr><td id="select">cell2-2</td></tr>' +
                       "<tr><td>cell3-1</td><td>cell3-2</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              '<tr><td rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                              "<tr><td>cell3-2</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust rowspan when spanned <tr>'s last child cell element is removed (removing middle row)");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td rowspan="3">cell1-1</td><td>cell1-2</td></tr>' +
                       '<tr><td id="select1">cell2-2</td></tr>' +
                       '<tr><td>cell3-1</td><td id="select2">cell3-2</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              '<tr><td rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust rowspan when spanned <tr>'s last child cell element and remove the last row");

  // colspan
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select" colspan="2">cell1-1</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should just remove the cell spanning columns");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1" colspan="2">cell1-1</td><td>cell1-3</td></tr>' +
                       '<tr><td>cell2-1</td><td id="select2">cell2-2</td><td>cell2-3</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should just remove the cell spanning columns and the other selected cell element");

  // XXX broken case, colspan is not adjusted.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td colspan="2">cell1-1</td><td>cell1-3</td></tr>' +
                       '<tr><td>cell2-1</td><td id="select">cell2-2</td><td>cell2-3</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              '<tr><td colspan="1">cell1-1</td><td>cell1-3</td></tr>' +
                              "<tr><td>cell2-1</td><td>cell2-3</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust different row's colspan when corresponding cell element is removed");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td colspan="2">cell1-1</td><td>cell1-3</td></tr>' +
                       '<tr><td>cell2-1</td><td id="select1">cell2-2</td><td id="select2">cell2-3</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  todo_is(editor.innerHTML, "<table><tbody>" +
                              '<tr><td colspan="1">cell1-1</td><td>cell1-3</td></tr>' +
                              "<tr><td>cell2-1</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust different row's colspan when corresponding cell element is removed (when 2 cell elements are selected)");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td>cell1-1</td><td id="select1">cell1-2</td><td>cell1-4</td></tr>' +
                       '<tr><td id="select2" colspan="2">cell2-1</td><td>cell2-3</td><td>cell2-4</td></tr>' +
                     "</table>";
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                              "<tr><td>cell1-1</td><td>cell1-4</td></tr>" +
                              "<tr><td>cell2-3</td><td>cell2-4</td></tr>" +
                            "</tbody></table>",
          "nsITableEditor.deleteTableCellContents(1) should adjust different row's colspan when corresponding cell element is removed (when 2 cell elements are selected)");

  // When a cell contains first selection range, it should be removed.
  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  editor.scrollTop; // requires layout information.
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().deleteTableCell(1);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-2</td><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(1) should remove only a cell containing first selection range when there is no selected cell element");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td><td>cell1-3</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                     "</table>";
  editor.scrollTop; // requires layout information.
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 1);
  getTableEditor().deleteTableCell(2);
  is(editor.innerHTML, "<table><tbody>" +
                         "<tr><td>cell1-3</td></tr>" +
                         "<tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents(2) should remove only 2 cell elements starting from a cell containing first selection range when there is no selected cell element");

  SimpleTest.finish();
});

function getTableEditor() {
  var editingSession = SpecialPowers.wrap(window).docShell.editingSession;
  return editingSession.getEditorForWindow(window).QueryInterface(SpecialPowers.Ci.nsITableEditor);
}

</script>
</body>

</html>
